+++ src/furamaResort/models/Villa.java
@@ -4,19 +4,19 @@ import java.io.Serializable;
 
 public class Villa extends Facility {
     private String standard;  // đơn lập, song lập, liền kề ?
-    private float poolArea;
+    private double poolArea;
     private int floors;
 
     public Villa() {
     }
 
-    public Villa(String standard, float poolArea, int floors) {
+    public Villa(String standard, double poolArea, int floors) {
         this.standard = standard;// tiêu chuẩn phòng
         this.poolArea = poolArea;
         this.floors = floors;
     }
 
-    public Villa(String serviceCode, float useArea, double rentalFees, int maxNumsPeople, String rentalType, String standard, float poolArea, int floors) {
+    public Villa(String serviceCode, double useArea, double rentalFees, int maxNumsPeople, String rentalType, String standard, double poolArea, int floors) {
         super(serviceCode, useArea, rentalFees, maxNumsPeople, rentalType);
         this.standard = standard;
         this.poolArea = poolArea;
@@ -32,11 +32,11 @@ public class Villa extends Facility {
         this.standard = standard;
     }
 
-    public float getPoolArea() {
+    public double getPoolArea() {
         return poolArea;
     }
 
-    public void setPoolArea(float poolArea) {
+    public void setPoolArea(double poolArea) {
         this.poolArea = poolArea;
     }
